Rename function app project
