+++ src/PowerBiProcessor/Properties/serviceDependencies.json
