+++ src/PowerBiProcessor/Properties/serviceDependencies.local.json
